+++ huracanpy/_data/__init__.py
@@ -2,7 +2,6 @@
 
 __all__ = [
     "load",
-    "load_list",
     "save",
     "example_csv_file",
     "example_parquet_file",
@@ -23,7 +22,7 @@ __all__ = [
 
 import pathlib
 
-from ._load import load, load_list
+from ._load import load
 from ._save import save
 
 here = pathlib.Path(__file__).parent